refactor: extract SidebarFactory from WebPageFactory
Moves 3 sidebar selector methods (~65 lines) into a dedicated
SidebarFactory with zero constructor dependencies.
